backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.3
alembic==1.13.1
python-dotenv==1.0.0

# Phase III AI/MCP dependencies
openai==1.54.0
dateparser==1.2.0
rapidfuzz==3.10.0

# Development dependencies
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.3
